renamed constants file to data where all the questions exist
